Update: routesへユーザ情報照合の役割を兼任させない ユーザ情報照合はLayoutへ切り離す
