+++ src/components/layout/Layout/Layout.module.scss
@@ -1,0 +1,7 @@
+@use "style/_variable" as *;
+
+.layout {
+  position: relative;
+  display: grid;
+}
+
